Add recursive makefiles to build tests from the 3.1.1 version of qrencode (see https://github.com/digdog/qrencode)
- Move tests to one directory per test program
- When compiling the tests, the libraries have to be built by passing
a compile flag to make:
make QRENCODER_CCFLAGS='-DBUILD_QRENCODER_TESTS'
This enables the tests to call functions that would otherwise be declared "static".
- Add debug library build for ARM and X86
- Change test script (test_all.sh) to take arguments specifying the
CPU and release type. Also creates a symbolic link to the expected name
of the shared library. Reports the number of tests that pass
and the total number.